[tune,autoscaler] Test yaml, add better distributed docs (#5403)
