+++ python/ray/tune/examples/tune-default.yaml
@@ -1,51 +1,10 @@
-# An unique identifier for the head node and workers of this cluster.
-cluster_name: tune-example
-
-# The minimum number of workers nodes to launch in addition to the head
-# node. This number should be >= 0.
-min_workers: 2
-
-# The maximum number of workers nodes to launch in addition to the head
-# node. This takes precedence over min_workers.
-max_workers: 2
-
-# Cloud-provider specific configuration.
-provider:
-    type: aws
-    region: us-west-2
-    # Availability zone(s), comma-separated, that nodes may be launched in.
-    # Nodes are currently spread between zones by a round-robin approach,
-    # however this implementation detail should not be relied upon.
-    availability_zone: us-west-2a,us-west-2b
-
-# How Ray will authenticate with newly launched nodes.
-# By default Ray creates a new private keypair, but you can also use your own.
-auth:
-    ssh_user: ubuntu
-
-# Provider-specific config for the head node, e.g. instance type.
-head_node:
-    InstanceType: c5.xlarge
-    ImageId: ami-0b294f219d14e6a82 # Deep Learning AMI (Ubuntu) Version 21.0
-
-# Provider-specific config for worker nodes, e.g. instance type.
-worker_nodes:
-    InstanceType: c5.xlarge
-    ImageId: ami-0b294f219d14e6a82 # Deep Learning AMI (Ubuntu) Version 21.0
-
-    # Run workers on spot by default. Comment this out to use on-demand.
-    InstanceMarketOptions:
-        MarketType: spot
-
-# Files or directories to copy to the head and worker nodes. The format is a
-# dictionary from REMOTE_PATH: LOCAL_PATH, e.g.
-file_mounts: {
-#    "/path1/on/remote/machine": "/path1/on/local/machine",
-#    "/path2/on/remote/machine": "/path2/on/local/machine",
-}
-
-# List of shell commands to run to set up each node.
-setup_commands:
-    - pip install -U https://s3-us-west-2.amazonaws.com/ray-wheels/latest/ray-0.8.0.dev3-cp36-cp36m-manylinux1_x86_64.whl
-    - pip install torch torchvision tabulate tensorboard filelock
-
+cluster_name: tune-default
+provider: {type: aws, region: us-west-2}
+auth: {ssh_user: ubuntu}
+min_workers: 3
+max_workers: 3
+# Deep Learning AMI (Ubuntu) Version 21.0
+head_node: {InstanceType: c5.xlarge, ImageId: ami-0b294f219d14e6a82}
+worker_nodes: {InstanceType: c5.xlarge, ImageId: ami-0b294f219d14e6a82}
+setup_commands: # Set up each node.
+    - pip install ray torch torchvision tabulate tensorboard
